+++ api/ai/external_algorithms/group_matcher_algorithm/utils.py
@@ -1,11 +1,30 @@
 from typing import List
 
-from api.ai.external_algorithms.group_matcher_algorithm.custom_modes import (
-    AlGender,
-    AlRace,
-    AlYearLevel,
-)
-from api.models.enums import Gender, Race
+from api.models.enums import Gender, Race, AttributeValueEnum
+
+
+class AlRace(AttributeValueEnum):
+    White = "White"
+    Asian = "Asian"
+    Hispanic = "Hispanic"
+    Black_Or_African_American = "Black/African American"
+    Indegenous = "Indegenous"
+    Middle_Eastern = "Middle-Eastern"
+    Multiple_Races = "Multiple races"
+
+
+class AlYearLevel(AttributeValueEnum):
+    Freshman = "freshman"
+    Sophomore = "sophomore"
+    Junior = "junior"
+    Senior = "senior"
+    Graduate = "graduate"
+
+
+class AlGender(AttributeValueEnum):
+    Female = "Female"
+    Male = "Male"
+    Other = "Other"
 
 
 def fromGenderToAlGender(gender: Gender) -> AlGender:
